updated home page
